+++ src/App.tsx
@@ -6,28 +6,29 @@ import "./index.css";
 import Homepage from "./components/pages/homepage";
 import AboutPage from "./components/pages/aboutpage";
 import ContactPage from "./components/pages/contactpage";
-import ServicePage from "./components/pages/servicepage";
 import InsightPage from "./components/pages/insightpage";
 import AdminPage from "./components/pages/adminpage";
 import ResetPasswordPage from "./components/pages/resetpasswordpage";
 import ResetSuccessPage from "./components/pages/resetsuccess";
 import BgImage from "./components/bg_image";
+import { AuthProvider } from "./lib/context/auth_context";
 
 const App: React.FC = () => {
 	return (
 		<Router>
-			<Sidebar />
-			<BgImage />
-			<Routes>
-				<Route path="/" element={<Homepage />} />
-				<Route path="/about" element={<AboutPage />} />
-				<Route path="/services" element={<ServicePage />} />
-				<Route path="/contact" element={<ContactPage />} />
-				<Route path="/admin" element={<AdminPage />} />
-				<Route path="/insight" element={<InsightPage />} />
-				<Route path="/reset-password" element={<ResetPasswordPage />} />
-				<Route path="/reset-success" element={<ResetSuccessPage />} />
-			</Routes>
+			<AuthProvider>
+				<Sidebar />
+				<BgImage />
+				<Routes>
+					<Route path="/" element={<Homepage />} />
+					<Route path="/about" element={<AboutPage />} />
+					<Route path="/contact" element={<ContactPage />} />
+					<Route path="/admin" element={<AdminPage />} />
+					<Route path="/insight" element={<InsightPage />} />
+					<Route path="/reset-password" element={<ResetPasswordPage />} />
+					<Route path="/reset-success" element={<ResetSuccessPage />} />
+				</Routes>
+			</AuthProvider>
 		</Router>
 	);
 };